requirements.txt
Flask==3.1.0
requests==2.32.3
pyttsx3==2.98
python-dotenv==1.0.1
openpyxl==3.1.5
